feat(api): move anonymize controller in src
